kiwijob/apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-multipart==0.0.17
pypdf==5.1.0
python-docx==1.1.2
openai==1.57.4
python-dotenv==1.0.1
httpx==0.28.1
pydantic-settings==2.6.1
eval-type-backport==0.2.2
google-auth==2.36.0
PyJWT[crypto]==2.10.1
pytest==8.3.4
